+++ vegitable/shops/shop_views/settings_view.py
@@ -1,5 +1,4 @@
-from django.shortcuts import redirect, render, get_object_or_404
-
+from django.shortcuts import render
 from ..models import Shop, Index
