+++ vegitable/template/Report/shilk.html
@@ -3,7 +3,7 @@
 {% block content %}
 {% include "menubar.html" %}
 <div class="container">
-    <h1 class="heading">Shilk Entry</h1>
+    <h1 class="heading">Shilk Report</h1>
     <div class="clearfix"> </div>
 
     <div class="shilk_container">
@@ -67,4 +67,5 @@
 
     <div class="clearfix"> </div>
 </div>
+<script src="{% static 'js/shilk_report/shilk_report.js'%}"></script>
 {% endblock %}
