+++ vegitable/template/menubar.html
@@ -36,16 +36,16 @@
         <li><a href="/patti_list">Patti Entry</a></li>
         <li><a href="/expenditure_entry">Expenditure Entry</a></li>
         <li><a href="/credit_bill_entry">CreditBill Entry</a></li>
-        <li><a href="/shilk_entry">Shilk Entry</a></li>
       </ul>
     </li>
     <li class="submenu">
       <a href="#" class="{% if 'report' in request.path %}active{% endif %}"
-        >Report</a
+        >Reports</a
       >
       <ul class="submenu-list">
         <li><a href="/sales_bill_report">Sales Report</a></li>
         <li><a href="/rmc_report">RMC Report</a></li>
+        <li><a href="/shilk_entry">Shilk Report</a></li>
       </ul>
     </li>
     <li class="submenu">
